add models.py
